Definitions:
Jump: An edge from level a to level b where (b-a) > 1

This is the first draft of edges without jumps. Boy, there are so many cases to handle. Much more than I had anticipated in the morning. Tomorrow, I'll tackle edges with jumps.

Also, some refactoring has to be done. As of now, I'm focusing on the logic and not so much on clean code.