Removed Versioning in Quiz Model  - Django Changes:
Changed the Server API to a 'normal' API where POST creates a Quiz and PUT modifies the Quiz. But how are we going to fix the concurrency issue? We're going to use a websocket between the client and a node server which aggregates messages from the client and calls the appropriate Django  API once per second.  All messages in a web socket will appear in the right order at the node web socket server. So we don't have to worry about ordering anymore.