WebSocket: First commit in implementing a more robust web socket for auto save. Details below.
We need to generate ids on the client to save data using WebSocket so that PUT requests can follow POST requests before the id is returned by the server. So auto incrementing integer ids are not feasible and we need to use UUIDs instead.

So I went down a rabbit hole and found that django-taggit does not support UUIDs. So I made the changes explained in my answer on StackOverflow

Also changed individual quiz urls from /api/v1/quiz to /api/v1/quizzes to make quiz urls consistent.