Concept API: Fixed JSON serialization issue in nested ConceptSection data.
Since we're using a TextField to store JSON data, rest_framework's ModelSerializer converts the Python dictionary which stores the JSON data into a string. But we want to store the dictionary using json.dumps(dict) instead of str(dict). So we are using the serializer for validation purposes and using the raw request.data to perform json.dumps.