Conceptviz DAG: In the topological sort routine, sort source nodes first
Previously, we are sorting from back to front, i.e. sink nodes were added to the last level. Prereqs to sink nodes were added to the penultimate level and so on. But this caused bad visualisation. A better method is to start from source nodes, and work front to back resulting in a more natal visualisation.